+++ package.json
@@ -17,6 +17,7 @@
     "react-dom": "^17.0.1",
     "react-google-button": "^0.7.1",
     "react-hooks-helper": "^1.6.0",
+    "react-infinite-scroll-component": "^6.0.0",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
